Relocate organization struct
