+++ auth/user.go
@@ -113,26 +113,9 @@ type UserOrganization struct {
 	Role           string `gorm:"default:'member'"`
 }
 
-// Organization struct
-type Organization struct {
-	ID        uint      `gorm:"primary_key" json:"id"`
-	GithubID  *int64    `gorm:"unique" json:"githubId,omitempty"`
-	CreatedAt time.Time `json:"createdAt"`
-	UpdatedAt time.Time `json:"updatedAt"`
-	Name      string    `gorm:"unique;not null" json:"name"`
-	Provider  string    `gorm:"not null" json:"provider"`
-	Users     []User    `gorm:"many2many:user_organizations" json:"users,omitempty"`
-	Role      string    `json:"-" gorm:"-"` // Used only internally
-}
-
 // IDString returns the ID as string
 func (user *User) IDString() string {
 	return fmt.Sprint(user.ID)
-}
-
-// IDString returns the ID as string
-func (org *Organization) IDString() string {
-	return fmt.Sprint(org.ID)
 }
 
 // TableName sets CICDUser's table name
